Don't mark the overridable attributes as O+C, but optional only (#25)
* Don't mark the overridable attributes as O+C, but optional only

We don't have to always set a value for these attributes (e.g. `create_method`), as the information needed for the API operations are always available. So if an attribute is only provided at the provider level, that attribute is used for CRUD of the resource/data source, but the value is not set to the resource/data source's state, which makes the code a lot cleaner.

Another benefit of that is in this case, if another attribute of that resource changed (e.g. the `body`), the plan diff will show unnecessary noise telling users that those O+C and not yet set attribute (so they are regarded as C at this moment) as "known after apply". This will be avoided with this change.

* Update tests