Support `precheck` (#26)
This PR add following supports for pre-check:

- `restful_resource` (resource): adds `precheck_{check|update|delete}`
- `restful_operation`: adds `precheck`

The precheck is similar to the `poll_xxx` attribute, except the precheck is polled before the actual operation. Another difference is that, instead of a `url_locator`, the precheck used the `path` and `query`.

Some useful use cases:

- [Some API](https://help.skytap.com/api-busy-bp.html) requires the caller to manual check condition prior to invoking a request
- Some resource will be put in a middle state when its child resource got created/updated/deleted (e.g. Azure firewall policy and its child rule collection groups), where in that middle state, only one of the child resources can be modified. Imagine there are two parallel workflows which reference the parent resource via a data source and create independent child resources at the same time, only one workflow can work, the other one will fail (e.g. due to another operation is under going). With the precheck, one can poll the status of the parent resource, and only proceed when it is in a ready status. Though this is not a solid solution to avoid the conflicting operations, whilst it can narrow down the conflicting window a lot.

Due to https://github.com/hashicorp/terraform-plugin-framework/issues/566, the precheck for the read operation of `restful_resource` resource and for the `restful_resource` data source is not implemented. Also, there isn't a clear use case for it.

Fix #24