`poll_xxx`: Fix double urlencoding the URL and introduce the `query` attribute
Previously, we were directly setting the `RawRequest.URL` as the encoded url. This causes the double encoding when we construct the pollable via calling the `url.String()`. Instead, we should set the encoded url to `URL`, and when constructing the pollable, we shall firstly `url.Parse()` the request URL. As a result, the `Path` will be in decoded form, and if the url is encoded before, the `RawPath` will be non-empty, which is a encode hint. Then, we intentionally clear the query part in the parsed URL, then call `url.String()` to get the encoded form and set it to the pollable's URL.

Since we are cleaning up the query string, we need to accordingly allow users to modify arbitrary query parameters so no to break their current use cases. Therefore, we introduce the `poll_xxx.query` attribute.